LPS-127106 Move isMapped to utils/editable-value
