+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/utils/editable-value/isMappedToStructure.js
@@ -12,9 +12,6 @@
  * details.
  */
 
-export default function isMapped(editable) {
-	return (
-		editable &&
-		(editable.fieldId || editable.mappedField || editable.collectionFieldId)
-	);
+export default function isMappedToStructure(editable) {
+	return editable && editable.mappedField;
 }
